requirements.txt
# Career-HY RAG 실험 파이프라인 의존성

# 기본 라이브러리
pyyaml>=6.0
boto3>=1.28.0
numpy>=1.24.0

# AI/ML 라이브러리
openai>=1.0.0
langchain>=0.1.0
chromadb>=0.4.0
tiktoken>=0.5.0

# PDF 처리
PyMuPDF>=1.23.0

# 평가 및 분석
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# 개발 도구
pytest>=7.0.0
black>=23.0.0
mypy>=1.5.0